Refactor: Refactor ApplicationService to Serviceable
