--- app/services/application_service/result/failure.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-class ApplicationService
-  class Result
-    class Failure < Result
-      def success?
-        false
-      end
-
-      def on_failure
-        yield(self)
-      end
-
-      def on_success; end
-    end
-  end
-end
