+++ app/services/serviceable.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Serviceable
+  extend ActiveSupport::Concern
+
+  include Resultable
+end
