+++ app/services/serviceable/resultable.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Serviceable
+  module Resultable
+    extend ActiveSupport::Concern
+
+    def success(metadata = {})
+      Result.success(metadata)
+    end
+
+    def failure(metadata = {})
+      Result.failure(metadata)
+    end
+  end
+end
